feat(dsv4): FP8 prefill forward path — partial wireup
Replaces the Phase 3 NotImplementedError guard with a working FP8
prefill forward built on top of upstream's ``_forward_body``. Reuses
existing FP8 helpers shipped with the rebased FP8 chain:

* SWA quant-write: ``_prefill_write_swa_to_pool`` FP8 branch routes
  bf16 ``kv_full`` through ``_swa_fp8_kv_insert_triton.quantize_and_insert_k_cache``
  (writes 584B/slot fp8 NoPE + bf16 RoPE + ue8m0 scale).
* CSA/HCA dequant: ``_prefill_paged_read_kv`` FP8 branch dequantizes
  via ``_swa_fp8_dequant_triton.dequantize_and_gather_k_cache``,
  returning bf16 ``[bsz, T, 512]`` so the existing BF16 sparse_attn
  dispatch (kv_cat path) is reused unchanged.
* Compressor pool round-trip: after ``CompressorFP8`` writes to the
  CSA/HCA pool and returns None, ``_forward_body`` materializes a
  bf16 ``kv_compress`` by reading [cmp_write_start, NB_total) back out
  via the FP8 dequant branch above. The BF16 ``_prefill_paged_write_kv_range``
  is skipped via ``_fp8_compressor_handled_pool_write`` to avoid
  corrupting the per-block striped FP8 layout with a bf16 overwrite.
* 3D pool view: new ``_pool_view_3d_fp8`` builds an as_strided
  ``[num_blocks, eb, ENTRY_BYTES]`` view honoring C++-side TMA per-
  block padding. Replaces the flat 2D ``_pool_view`` (which crashes
  with ``view size is not compatible with input tensor's size and
  stride`` for FP8 pools that have padding gaps). Wired into both
  ``_set_compressor_pool_context`` (passing 3D view to CompressorFP8 /
  IndexerFP8) and ``_prefill_paged_read_kv``.
* ``Attention.forward`` accepts ``attn_inputs=None`` kwarg so the
  block.py call site (which now passes attn_inputs from upstream's
  framework descriptor wiring) stops raising TypeError.
* CompressorFP8 keeps wkv/wgate as fp32 — the FP8 compressor body
  accumulates partials in fp32; the BF16 cast that BF16 Compressor
  uses is body-internal and not portable across paths.

Status: BF16 path remains byte-equal vs pre-rebase (verified). FP8
forward end-to-end NOT yet validated — last server boot failed at
``compressor_fp8.py:314`` with a residual fp32/bf16 matmul mismatch
that this commit's wkv/wgate fp32 fix should resolve; needs another
boot iteration to confirm.

+++ rtp_llm/models_py/modules/dsv4/compressor_fp8.py
@@ -114,17 +114,20 @@ class CompressorFP8(PoolBackedModule):
         self.rotate = rotate
         coff = 1 + self.overlap
 
-        # ape stays FP32; wkv/wgate cast BF16 for tensor-core dispatch
-        # (matches BF16 ``Compressor`` ctor pattern).
-        self.ape = compressor_weights["ape"]
+        # ape / wkv / wgate stay FP32 — CompressorFP8._forward_prefill_body
+        # accumulates in FP32 for the (kv * softmax(score)).sum reduction
+        # before the FP8 quant write. (BF16 ``Compressor`` casts wkv/wgate
+        # to BF16 because its body keeps the partials in BF16; FP8 path
+        # diverges here.)
+        self.ape = compressor_weights["ape"].float()
         self.wkv = nn.Linear(dim, coff * head_dim, bias=False)
         self.wgate = nn.Linear(dim, coff * head_dim, bias=False)
         with torch.no_grad():
             self.wkv.weight = nn.Parameter(
-                compressor_weights["wkv"].to(torch.bfloat16), requires_grad=False
+                compressor_weights["wkv"].float(), requires_grad=False
             )
             self.wgate.weight = nn.Parameter(
-                compressor_weights["wgate"].to(torch.bfloat16), requires_grad=False
+                compressor_weights["wgate"].float(), requires_grad=False
             )
         self.norm = _CompressorNorm(head_dim)
         self.norm.weight = nn.Parameter(
